Fix stack overflow on Marshal
We use a method for certain structures to marshal it. I'm not sure why,
since it has virtually no difference with `{json, proto}.Marshal()`, but
we do.

Anyway, protobuffer does the same internally. When we overwrite the
method we created an infinite call loop. This fixes it by removing the
unnecessary method.