Issue #55, #60: int -> int64
Why not `uint64`? Because that requires more non trivial changes, that
fail in ways I still can't understand.

Basically, the code uses the assumption that "undefined" is -1 and
"defined" is always bigger than -1. This let's you do comparisons to
check validity such as "id1 > id2."

If I change to uint64, I have to trace back each of those places and
change the logic. This is just a mere refactor, whose whole purpose it's
to allow to continue working in the protobuffer transition.