int64* -> int64
Change all the references that use pointer to `int64` to use pure
`int64`. This is required to make the transition to protobuffer, as
there's no such a concept as pointers in the messages schema,

- Fix funky test

This test stopped passing after merging with master. That's because the
test it's poorly designed.

In production, it passes because when the event isnt' found in the
store, it returns in the first argument a newly created, empty event
which happens to have `roundReceived` initialized to `nil`. When using
integers instead of pointers, the value is 0 and not -1, the equivalent
we're using. Instead, let's ignore all events that aren't in the store.

Terrible solution, but it's a terrible test anyways.