Issue #55: Second step to protobuffers
Sending pointers through protobuffer it's not possible. That's a
problem, because one of the structures we have to transmit, `Event`, has
some internal private state saved as pointers.

Initially, I tried to change it to plain integers. That failed
misserably for misterious (and unknown) reasons.

So I decided to take another approache: Apply good OOP practices and
divide `Event` in two. One, the wrapper, contains the internal state and
the message to be marshelled. The message, contains only the fields that
will be transmitted through the network.

That way, we don't have to marshall anything with pointers.